+++ CHANGELOG.md
@@ -28,6 +28,7 @@ branch is the latest unreleased version.
 - Fix Instance.FindFirstAncestor behaving as FindFirstAncestorOfClass.
 - Fix equality of Enums, Enum, and EnumItem types.
 - Fix tostring of Enums, Enum, and EnumItem types.
+- Fix instance properties not checking inherited classes.
 
 **Internal:**
 - Automated tests run on Windows in addition to Linux.